Move ShoppingList classes to own package
