+++ app/src/main/java/kres/realtimeshoppinglist/firebase/shoppingList/ShoppingListNameChangeListener.java
@@ -1,4 +1,4 @@
-package kres.realtimeshoppinglist.firebase;
+package kres.realtimeshoppinglist.firebase.shoppingList;
 
 import android.util.Log;
